TEST: backend bench增加参数

+++ benchmark/test_redis_backend.py
@@ -2,6 +2,7 @@ import asyncio
 import time
 import docker
 import numpy as np
+import argparse
 
 from hetu.data import (
     define_component, Property, BaseComponent,
@@ -37,15 +38,14 @@ async def timeit(func, repeat=1, repeat_mul=1, concurrency=100, *args):
     return cost, repeat * repeat_mul / cost
 
 
-async def run_bench():
-    backend = RedisBackendClientPool({"master": "redis://127.0.0.1:23318/0"})
-    item_data = RedisComponentBackend(Item, 'test', 1, backend)
+async def run_bench(inst, redis_address):
+    backend = RedisBackendClientPool({"master": redis_address})
+    item_data = RedisComponentBackend(Item, inst, 1, backend)
     # clean db
-    async with item_data.transaction() as _tbl:
-        _rows = await _tbl.query('id', -np.inf, +np.inf)
-        for _row in _rows:
-            await _tbl.delete(_row.id)
-        print(f'clean {len(_rows)} rows')
+    keys = backend.io.keys(f'{item_data._root_prefix}*')
+    if keys:
+        backend.io.delete(*keys)
+    print(f'clean {len(keys)} rows')
 
     import os
     print("pid:", os.getpid())
@@ -67,21 +67,71 @@ async def run_bench():
                 continue
             return retry
 
-    t, qps = await timeit(test_insert, 30000, 1)
+    async def test_select(i):
+        retry = 0
+        while True:
+            try:
+                async with item_data.transaction() as tbl:
+                    row = await tbl.select(i+1)
+            except RaceCondition as e:
+                print(e)
+                retry += 1
+                continue
+            return retry
+
+    async def test_update(i):
+        retry = 0
+        while True:
+            try:
+                async with item_data.transaction() as tbl:
+                    row = await tbl.select(i+1)
+                    row.name = f'Itm{i}'
+                    await tbl.update(row.id, row)
+            except RaceCondition as e:
+                print(e)
+                retry += 1
+                continue
+            return retry
+
+    async def test_direct_update(i):
+        name = await backend.aio.hget(f'{item_data._key_prefix}{i+1}', 'name')
+        await backend.aio.hset(f'{item_data._key_prefix}{i+1}', 'name', f'It2{i}')
+        return 0
+
+    t, qps = await timeit(test_insert, 3000, 1)
     # 单worker 1000/s
+    assert qps >= 500, 'benchmark redis太慢，检查下'
+
+    t, qps = await timeit(test_select, 3000, 1)
+    # 单worker pipeline get 2000/s
+
+    t, qps = await timeit(test_update, 3000, 1)
+    # 单worker 1000/s
+
+    t, qps = await timeit(test_direct_update, 6000, 1)
+    # 单worker 3000/s
 
 
 if __name__ == '__main__':
-    # 启动local redis docker
-    client = docker.from_env()
-    try:
-        client.containers.get('hetu_test_redis').kill()
-        client.containers.get('hetu_test_redis').remove()
-    except (docker.errors.NotFound, docker.errors.APIError):
-        pass
-    container = client.containers.run(
-        "redis:latest", detach=True, ports={'6379/tcp': 23318},
-        name='hetu_test_redis', auto_remove=True)
+    parser = argparse.ArgumentParser(prog='hetu', description='Hetu Data Server')
+    parser.add_argument("--instance_name", default="test")
+    parser.add_argument("--address")
+    args = parser.parse_args()
 
-    asyncio.run(run_bench())
-    container.kill()
+    # 启动local redis docker
+    address = args.address
+    if args.address is None:
+        address = 'redis://127.0.0.1:23318/0'
+        client = docker.from_env()
+        try:
+            client.containers.get('hetu_test_redis').kill()
+            client.containers.get('hetu_test_redis').remove()
+        except (docker.errors.NotFound, docker.errors.APIError):
+            pass
+        container = client.containers.run(
+            "redis:latest", detach=True, ports={'6379/tcp': 23318},
+            name='hetu_test_redis', auto_remove=True)
+
+    asyncio.run(run_bench(args.instance_name, address))
+    if args.address is None:
+        container.kill()
